Move all persistent subscription consumer strategy related classes to a new namespace.
