+++ src/EventStore.Core/EventStore.Core.csproj
@@ -184,10 +184,15 @@
     <Compile Include="Services\Gossip\LoopbackDns.cs" />
     <Compile Include="Services\Gossip\NodeGossipService.cs" />
     <Compile Include="Services\Histograms\HistogramService.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\ConsumerPushResult.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\DelegatePersistentSubscriptionConsumerStrategyFactory.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\DispatchToSinglePersistentSubscriptionConsumerStrategy.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\IPersistentSubscriptionConsumerStrategy.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\IPersistentSubscriptionConsumerStrategyFactory.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\PersistentSubscriptionConsumerStrategyRegistry.cs" />
+    <Compile Include="Services\PersistentSubscription\ConsumerStrategy\RoundRobinPersistentSubscriptionConsumerStrategy.cs" />
     <Compile Include="Services\PersistentSubscription\IPersistentSubscriptionCheckpointReader.cs" />
     <Compile Include="Services\PersistentSubscription\IPersistentSubscriptionCheckpointWriter.cs" />
-    <Compile Include="Services\PersistentSubscription\IPersistentSubscriptionConsumerStrategy.cs" />
-    <Compile Include="Services\PersistentSubscription\IPersistentSubscriptionConsumerStrategyFactory.cs" />
     <Compile Include="Services\PersistentSubscription\IPersistentSubscriptionMessageParker.cs" />
     <Compile Include="Services\PersistentSubscription\IPersistentSubscriptionStreamReader.cs" />
     <Compile Include="Services\PersistentSubscription\NakAction.cs" />
@@ -199,7 +204,6 @@
     <Compile Include="Services\PersistentSubscription\PersistentSubscriptionClient.cs" />
     <Compile Include="Services\PersistentSubscription\PersistentSubscriptionClientCollection.cs" />
     <Compile Include="Services\PersistentSubscription\PersistentSubscriptionConfig.cs" />
-    <Compile Include="Services\PersistentSubscription\PersistentSubscriptionConsumerStrategyRegistry.cs" />
     <Compile Include="Services\PersistentSubscription\PersistentSubscriptionMessageParker.cs" />
     <Compile Include="Services\PersistentSubscription\PersistentSubscriptionParams.cs" />
     <Compile Include="Services\PersistentSubscription\PersistentSubscriptionParamsBuilder.cs" />
