+++ src/EventStore.Core/PluginModel/IPersistentSubscriptionConsumerStrategyPlugin.cs
@@ -1,4 +1,5 @@
 ﻿using EventStore.Core.Services.PersistentSubscription;
+using EventStore.Core.Services.PersistentSubscription.ConsumerStrategy;
 
 namespace EventStore.Core.PluginModel
 {
