+++ src/EventStore.Core/Services/PersistentSubscription/ConsumerStrategy/ConsumerPushResult.cs
@@ -1,0 +1,9 @@
+namespace EventStore.Core.Services.PersistentSubscription.ConsumerStrategy
+{
+    public enum ConsumerPushResult
+    {
+        Sent,
+        Skipped,
+        NoMoreCapacity
+    }
+}
