+++ src/EventStore.Core/Services/PersistentSubscription/ConsumerStrategy/DelegatePersistentSubscriptionConsumerStrategyFactory.cs
@@ -1,15 +1,8 @@
 using System;
 using EventStore.Core.Bus;
 
-namespace EventStore.Core.Services.PersistentSubscription
+namespace EventStore.Core.Services.PersistentSubscription.ConsumerStrategy
 {
-    public interface IPersistentSubscriptionConsumerStrategyFactory
-    {
-        string StrategyName { get; }
-
-        IPersistentSubscriptionConsumerStrategy Create(string subscriptionId, IPublisher mainQueue, ISubscriber mainBus);
-    }
-
     class DelegatePersistentSubscriptionConsumerStrategyFactory : IPersistentSubscriptionConsumerStrategyFactory
     {
         public string StrategyName { get; }
