+++ src/EventStore.Core/Services/PersistentSubscription/ConsumerStrategy/IPersistentSubscriptionConsumerStrategy.cs
@@ -1,0 +1,15 @@
+using EventStore.Core.Data;
+
+namespace EventStore.Core.Services.PersistentSubscription.ConsumerStrategy
+{
+    public interface IPersistentSubscriptionConsumerStrategy
+    {
+        string Name { get; }
+
+        void ClientAdded(PersistentSubscriptionClient client);
+
+        void ClientRemoved(PersistentSubscriptionClient client);
+
+        ConsumerPushResult PushMessageToClient(ResolvedEvent ev);
+    }
+}
