+++ src/EventStore.Core/Services/PersistentSubscription/ConsumerStrategy/IPersistentSubscriptionConsumerStrategyFactory.cs
@@ -1,0 +1,11 @@
+using EventStore.Core.Bus;
+
+namespace EventStore.Core.Services.PersistentSubscription.ConsumerStrategy
+{
+    public interface IPersistentSubscriptionConsumerStrategyFactory
+    {
+        string StrategyName { get; }
+
+        IPersistentSubscriptionConsumerStrategy Create(string subscriptionId, IPublisher mainQueue, ISubscriber mainBus);
+    }
+}
