+++ src/EventStore.Core/Services/PersistentSubscription/ConsumerStrategy/RoundRobinPersistentSubscriptionConsumerStrategy.cs
@@ -3,49 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using EventStore.Core.Data;
 
-namespace EventStore.Core.Services.PersistentSubscription
+namespace EventStore.Core.Services.PersistentSubscription.ConsumerStrategy
 {
-    public enum ConsumerPushResult
-    {
-        Sent,
-        Skipped,
-        NoMoreCapacity
-    }
-
-    public interface IPersistentSubscriptionConsumerStrategy
-    {
-        string Name { get; }
-
-        void ClientAdded(PersistentSubscriptionClient client);
-
-        void ClientRemoved(PersistentSubscriptionClient client);
-
-        ConsumerPushResult PushMessageToClient(ResolvedEvent ev);
-    }
-
-    class DispatchToSinglePersistentSubscriptionConsumerStrategy : RoundRobinPersistentSubscriptionConsumerStrategy
-    {
-        public override string Name
-        {
-            get { return SystemConsumerStrategies.DispatchToSingle; }
-        }
-
-        public override ConsumerPushResult PushMessageToClient(ResolvedEvent ev)
-        {
-            for (int i = 0; i < Clients.Count; i++)
-            {
-                if (Clients.Peek().Push(ev))
-                {
-                    return ConsumerPushResult.Sent;
-                }
-                var c = Clients.Dequeue();
-                Clients.Enqueue(c);
-            }
-
-            return ConsumerPushResult.NoMoreCapacity;
-        }
-    }
-
     class RoundRobinPersistentSubscriptionConsumerStrategy : IPersistentSubscriptionConsumerStrategy
     {
         protected readonly Queue<PersistentSubscriptionClient> Clients = new Queue<PersistentSubscriptionClient>();
@@ -91,5 +50,4 @@ namespace EventStore.Core.Services.PersistentSubscription
             return ConsumerPushResult.NoMoreCapacity;
         }
     }
-
 }
